requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.9.1.post1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
bilibili-api-python==14.1.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
docopt==0.6.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
loguru==0.6.0
lxml==4.9.2
multidict==6.0.4
Pillow==9.4.0
pipreqs==0.4.11
pyasn1==0.4.8
PyMySQL==1.0.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.4.2
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
win32-setctime==1.1.0
yarg==0.1.9
yarl==1.8.2
